Add http interceptor
