+++ src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { ArticleModule } from './article/article.module';
@@ -20,7 +21,8 @@ import {
   ProfilesService,
   SharedModule,
   TagsService,
-  UserService
+  UserService,
+  HttpTokenInterceptor
 } from './shared';
 
 const rootRouting: ModuleWithProviders = RouterModule.forRoot([]);
@@ -43,6 +45,7 @@ const rootRouting: ModuleWithProviders = RouterModule.forRoot([]);
     SettingsModule
   ],
   providers: [
+    { provide: HTTP_INTERCEPTORS, useClass: HttpTokenInterceptor, multi: true},
     ApiService,
     ArticlesService,
     AuthGuard,
