+++ src/app/shared/index.ts
@@ -6,3 +6,4 @@ export * from './models';
 export * from './services';
 export * from './shared.module';
 export * from './show-authed.directive';
+export * from './interceptors';
