+++ src/app/shared/interceptors/index.ts
@@ -1,0 +1,1 @@
+export * from './http.token.interceptor';
