+++ src/app/shared/services/api.service.ts
@@ -14,47 +14,32 @@ export class ApiService {
     private jwtService: JwtService
   ) {}
 
-  private setHeaders(): HttpHeaders {
-    const headersConfig = {
-      'Content-Type': 'application/json',
-      'Accept': 'application/json'
-    };
-
-    if (this.jwtService.getToken()) {
-      headersConfig['Authorization'] = `Token ${this.jwtService.getToken()}`;
-    }
-    return new HttpHeaders(headersConfig);
-  }
-
   private formatErrors(error: any) {
     return new ErrorObservable(error.json());
   }
 
   get(path: string, params: HttpParams = new HttpParams()): Observable<any> {
-    return this.http.get(`${environment.api_url}${path}`, { headers: this.setHeaders(), params })
+    return this.http.get(`${environment.api_url}${path}`, { params })
       .pipe(catchError(this.formatErrors));
   }
 
   put(path: string, body: Object = {}): Observable<any> {
     return this.http.put(
       `${environment.api_url}${path}`,
-      JSON.stringify(body),
-      { headers: this.setHeaders() }
+      JSON.stringify(body)
     ).pipe(catchError(this.formatErrors));
   }
 
   post(path: string, body: Object = {}): Observable<any> {
     return this.http.post(
       `${environment.api_url}${path}`,
-      JSON.stringify(body),
-      { headers: this.setHeaders() }
+      JSON.stringify(body)
     ).pipe(catchError(this.formatErrors));
   }
 
   delete(path): Observable<any> {
     return this.http.delete(
-      `${environment.api_url}${path}`,
-      { headers: this.setHeaders() }
+      `${environment.api_url}${path}`
     ).pipe(catchError(this.formatErrors));
   }
 }
